# mobile/openapi/mobile-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InterventionRequestCreateRequest": {
        "description": "What's sent to create a brand-new intervention request.\n\nrequest_number and submitted_at are never accepted here \u2014 both are\ncomputed by the endpoint itself (see create_intervention_request).",
        "properties": {
          "cart_number": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cart Number"
          },
          "delivery_location": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Location"
          },
          "employee_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Name"
          },
          "employee_phone": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Phone"
          },
          "handled_by": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled By"
          },
          "preferred_delivery_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Delivery At"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "status_id": {
            "title": "Status Id",
            "type": "integer"
          },
          "team_cart_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Cart User Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "team_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          },
          "zone": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "question",
          "status_id"
        ],
        "title": "InterventionRequestCreateRequest",
        "type": "object"
      },
      "InterventionRequestResponse": {
        "description": "One intervention request, as shown on the Intervention Requests screen.",
        "properties": {
          "cart_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cart Number"
          },
          "delivery_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Location"
          },
          "employee_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Name"
          },
          "employee_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Phone"
          },
          "handled_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled By"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "preferred_delivery_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Delivery At"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "request_number": {
            "title": "Request Number",
            "type": "string"
          },
          "status_id": {
            "title": "Status Id",
            "type": "integer"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "team_cart_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Cart User Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "team_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "id",
          "request_number",
          "submitted_at",
          "team_id",
          "team_name",
          "cart_number",
          "question",
          "employee_name",
          "employee_phone",
          "preferred_delivery_at",
          "delivery_location",
          "zone",
          "status_id",
          "handled_by",
          "team_cart_user_id"
        ],
        "title": "InterventionRequestResponse",
        "type": "object"
      },
      "InterventionRequestUpdateRequest": {
        "description": "What's sent to update an existing intervention request \u2014 same shape\nas creating one (request_number/submitted_at stay fixed once set).",
        "properties": {
          "cart_number": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cart Number"
          },
          "delivery_location": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Location"
          },
          "employee_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Name"
          },
          "employee_phone": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Phone"
          },
          "handled_by": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handled By"
          },
          "preferred_delivery_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Delivery At"
          },
          "question": {
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "status_id": {
            "title": "Status Id",
            "type": "integer"
          },
          "team_cart_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Cart User Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "team_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Name"
          },
          "zone": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          }
        },
        "required": [
          "question",
          "status_id"
        ],
        "title": "InterventionRequestUpdateRequest",
        "type": "object"
      },
      "InterventionRequestsDashboardResponse": {
        "description": "Aggregate KPI stats for Intervention Requests' landing dashboard\n(\"KPI overview\").",
        "properties": {
          "closed_requests": {
            "title": "Closed Requests",
            "type": "integer"
          },
          "open_requests": {
            "title": "Open Requests",
            "type": "integer"
          },
          "status_breakdown": {
            "items": {
              "$ref": "#/components/schemas/InterventionRequestsStatusBreakdownItem"
            },
            "title": "Status Breakdown",
            "type": "array"
          },
          "team_breakdown": {
            "items": {
              "$ref": "#/components/schemas/InterventionRequestsTeamBreakdownItem"
            },
            "title": "Team Breakdown",
            "type": "array"
          },
          "total_requests": {
            "title": "Total Requests",
            "type": "integer"
          }
        },
        "required": [
          "total_requests",
          "open_requests",
          "closed_requests",
          "status_breakdown",
          "team_breakdown"
        ],
        "title": "InterventionRequestsDashboardResponse",
        "type": "object"
      },
...
      "InterventionRequestsTeamBreakdownItem": {
        "description": "One team/association name, with how many intervention requests were\nlogged for it \u2014 one bar of the KPI dashboard's team breakdown chart.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "team_name": {
            "title": "Team Name",
            "type": "string"
          }
        },
        "required": [
          "team_name",
          "count"
        ],
        "title": "InterventionRequestsTeamBreakdownItem",
        "type": "object"
      },
...
      "MobileModule2PermissionsResponse": {
        "description": "What the calling user may do in the phone's KarTracker: KarScan (the\nweb's \"Manuele kar beweging\"), Kar Planning and Kar Map. There is no\ndelete flag: deleting a movement is web-only.",
        "properties": {
          "can_create_actions": {
            "title": "Can Create Actions",
            "type": "boolean"
          },
          "can_view_actions": {
            "title": "Can View Actions",
            "type": "boolean"
          },
          "can_view_karmap": {
            "title": "Can View Karmap",
            "type": "boolean"
          },
          "can_view_karplanning": {
            "title": "Can View Karplanning",
            "type": "boolean"
          }
        },
        "required": [
          "can_view_actions",
          "can_create_actions",
          "can_view_karplanning",
          "can_view_karmap"
        ],
        "title": "MobileModule2PermissionsResponse",
        "type": "object"
      },
      "MobileModule3LookupsResponse": {
        "description": "Every dropdown list the request form needs, in one round trip.",
        "properties": {
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/InterventionStatusResponse"
            },
            "title": "Statuses",
            "type": "array"
          },
          "teamkar_members": {
            "items": {
              "$ref": "#/components/schemas/TeamKarMemberOptionResponse"
            },
            "title": "Teamkar Members",
            "type": "array"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/InterventionRequestsTeamResponse"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "statuses",
          "teams",
          "teamkar_members"
        ],
        "title": "MobileModule3LookupsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "RW Crew mobile API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mobile/v1/module-2/permissions": {
      "get": {
        "description": "What this user may do in the phone's KarTracker, so the app only\nshows the menu items and buttons the web roles allow.",
        "operationId": "get_my_permissions_api_mobile_v1_module_2_permissions_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileModule2PermissionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Permissions",
        "tags": [
          "mobile-module-2"
        ]
      }
    },
...
    "/api/mobile/v1/module-3/dashboard": {
      "get": {
        "description": "The KPI overview numbers, gated only by module access like on the web.",
        "operationId": "get_dashboard_api_mobile_v1_module_3_dashboard_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterventionRequestsDashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
    "/api/mobile/v1/module-3/intervention-requests": {
      "get": {
        "description": "Every request, newest first; open_only=true keeps only requests\nwhose status is an \"open\" one (the web screen's default filter).",
        "operationId": "list_intervention_requests_api_mobile_v1_module_3_intervention_requests_get",
        "parameters": [
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Open Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InterventionRequestResponse"
                  },
                  "title": "Response List Intervention Requests Api Mobile V1 Module 3 Intervention Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Intervention Requests",
        "tags": [
          "mobile-module-3"
        ]
      },
      "post": {
        "description": "Create a request. Its number and timestamp are set here, never by the app.",
        "operationId": "create_intervention_request_api_mobile_v1_module_3_intervention_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionRequestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterventionRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Intervention Request",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
    "/api/mobile/v1/module-3/intervention-requests/{request_id}": {
      "get": {
        "description": "One request, for the phone's detail/edit screen.",
        "operationId": "get_intervention_request_api_mobile_v1_module_3_intervention_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterventionRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intervention Request",
        "tags": [
          "mobile-module-3"
        ]
      },
      "put": {
        "description": "Update every editable field (number and timestamp stay fixed).",
        "operationId": "update_intervention_request_api_mobile_v1_module_3_intervention_requests__request_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterventionRequestUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterventionRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Intervention Request",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
    "/api/mobile/v1/module-3/lookups": {
      "get": {
        "description": "The statuses, teams and TeamKar members for the request form.\n\nGated by the requests screen's own permission (not the MasterData\n\"statuses\" screen), so anyone who can see requests can always name and\npick a status, exactly like the web's teams/TeamKar dropdown endpoints.",
        "operationId": "get_lookups_api_mobile_v1_module_3_lookups_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileModule3LookupsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lookups",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
    "/api/mobile/v1/module-3/permissions": {
      "get": {
        "description": "What this user may do on the requests screen (view / create / edit).",
        "operationId": "get_my_permissions_api_mobile_v1_module_3_permissions_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileModule3PermissionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Permissions",
        "tags": [
          "mobile-module-3"
        ]
      }
    },
...
  }
}